Consolidate VMThread
