+++ ddprof-lib/src/main/cpp/vmStructs.h
@@ -13,6 +13,7 @@
 #include <type_traits>
 #include "codeCache.h"
 #include "safeAccess.h"
+#include "threadState.h"
 
 namespace ddprof {
     class HeapUsage;
@@ -419,7 +420,23 @@ class JavaFrameAnchor : VMStructs {
     }
 };
 
-class VMThread : public /* TODO make private when consolidating VMThread? */ VMStructs {
+// Copied from JDK's globalDefinitions.hpp 'JavaThreadState' enum
+enum JVMJavaThreadState {
+    _thread_uninitialized     =  0, // should never happen (missing initialization)
+    _thread_new               =  2, // just starting up, i.e., in process of being initialized
+    _thread_new_trans         =  3, // corresponding transition state (not used, included for completeness)
+    _thread_in_native         =  4, // running in native code
+    _thread_in_native_trans   =  5, // corresponding transition state
+    _thread_in_vm             =  6, // running in VM
+    _thread_in_vm_trans       =  7, // corresponding transition state
+    _thread_in_Java           =  8, // running in Java or in stub code
+    _thread_in_Java_trans     =  9, // corresponding transition state (not used, included for completeness)
+    _thread_blocked           = 10, // blocked in vm
+    _thread_blocked_trans     = 11, // corresponding transition state
+    _thread_max_state         = 12  // maximum thread state+1 - used for statistics allocation
+};
+
+class VMThread : VMStructs {
   public:
     static VMThread* current();
 
@@ -455,9 +472,9 @@ class VMThread : public /* TODO make private when consolidating VMThread? */ VMS
                (vtbl[5] == _java_thread_vtbl[5]) >= 2;
     }
 
-    int state() {
-        return _thread_state_offset >= 0 ? SafeAccess::load32((int32_t*) at(_thread_state_offset), 0) : 0;
-    }
+    OSThreadState osThreadState();
+
+    int state();
 
     bool inJava() {
         return state() == 8;
